+++ co-kkiri/src/components/modals/EditUserProfileModal/ModalTextFieldDropdown.tsx
@@ -3,8 +3,8 @@ import MultiselectDropdown from "@/components/commons/DropDowns/StackMultiselect
 import { DROPDOWN_INFO } from "@/constants/dropDown";
 import { LABELS, ModalTextFieldInputConfig, REQUIRED, RULES } from "@/constants/textInputRules";
 import DESIGN_TOKEN from "@/styles/tokens";
-import { UserProfile } from "@/types/UserTypes";
-import { findOptionByValue } from "@/utils/ArrayUtils";
+import { UserProfile } from "@/types/userTypes";
+import { findOptionByValue } from "@/utils/arrayUtils";
 import { Control, Controller, ControllerRenderProps } from "react-hook-form";
 import styled from "styled-components";
