+++ co-kkiri/src/types/UserTypes.ts
@@ -2,7 +2,7 @@ export interface UserProfile {
   nickname: string;
   profileImageUrl: string;
   position: string;
-  career: number | undefined;
+  career: number | null;
   introduce: string;
   stack: string[];
   link: string;
